feat: modularize main launch pipeline and add cooperative cancel
Made-with: Cursor
