+++ src/renderer/layout/MainLayout.tsx
@@ -1323,185 +1323,36 @@ export default function App() {
     stopLaunchStatusPolling();
   }, [stopLaunchStatusPolling]);
 
-  const handleLaunch = async () => {
+  const handleCancelLaunch = useCallback(async () => {
+    const ctx = activeLaunchRunRef.current;
+    if (!ctx?.profileId || !ctx?.runId || !window.electron?.cancelProfileLaunch) return;
+    await window.electron.cancelProfileLaunch(ctx.profileId, ctx.runId);
+  }, []);
+
+  const handleLaunch = () => {
     if (!currentProfile?.id || !window.electron?.launchProfile) return;
     const launchProfileId = currentProfile.id;
-    stopLaunchStatusPolling();
-    activeLaunchRunRef.current = null;
-
-    if (launchFeedbackTimeoutRef.current) {
-      window.clearTimeout(launchFeedbackTimeoutRef.current);
-      launchFeedbackTimeoutRef.current = null;
-    }
-
-    setIsLaunching(true);
-    setLaunchFeedback({
-      status: "in-progress",
-      message: "Launching profile...",
-    });
-
-    let persistLaunchFeedback = false;
-    try {
-      if (window.electron?.saveProfiles) {
-        const serializableProfiles = toSerializableProfiles(
-          profiles,
-        );
-        const saveResult = await window.electron.saveProfiles(
-          serializableProfiles,
-        );
-        if (!saveResult?.ok) {
-          setLaunchFeedback({
-            status: "error",
-            message: "Could not save profile changes before launch.",
-          });
-          setIsLaunching(false);
-          launchFeedbackTimeoutRef.current = window.setTimeout(() => {
-            setLaunchFeedback({
-              status: "idle",
-              message: "",
-            });
-            launchFeedbackTimeoutRef.current = null;
-          }, 7000);
-          return;
-        }
-      }
-
-      const launchResult = await window.electron.launchProfile(
-        launchProfileId,
-      );
-      const launchRunId = String(launchResult?.runId || "").trim();
-      if (launchRunId) {
-        activeLaunchRunRef.current = {
-          profileId: launchProfileId,
-          runId: launchRunId,
-        };
-      }
-      const launchedApps = Number(launchResult?.launchedAppCount || 0);
-      const launchedTabs = Number(launchResult?.launchedTabCount || 0);
-      const failedCount = Array.isArray(launchResult?.failedApps)
-        ? launchResult.failedApps.length
-        : 0;
-      const skippedCount = Array.isArray(launchResult?.skippedApps)
-        ? launchResult.skippedApps.length
-        : 0;
-      const unresolvedPendingCount = Number(
-        launchResult?.unresolvedPendingConfirmationCount
-          || (Array.isArray(launchResult?.pendingConfirmations)
-            ? launchResult.pendingConfirmations.filter(
-              (item) => String(item?.status || "waiting").toLowerCase() !== "resolved",
-            ).length
-            : 0),
-      );
-      const pendingNames = Array.isArray(launchResult?.pendingConfirmations)
-        ? launchResult.pendingConfirmations
-          .map((item) => String(item?.name || "").trim())
-          .filter(Boolean)
-          .slice(0, 3)
-        : [];
-
-      if (launchResult?.ok) {
-        const summaryParts = [
-          `${launchedApps} app${launchedApps === 1 ? "" : "s"}`,
-          `${launchedTabs} tab${launchedTabs === 1 ? "" : "s"}`,
-        ];
-        if (failedCount > 0) summaryParts.push(`${failedCount} failed`);
-        if (skippedCount > 0) summaryParts.push(`${skippedCount} skipped`);
-        if (unresolvedPendingCount > 0) {
-          const canPollLaunchStatus = Boolean(launchRunId);
-          persistLaunchFeedback = canPollLaunchStatus;
-          const pendingList = pendingNames.length > 0
-            ? ` (${pendingNames.join(", ")}${unresolvedPendingCount > pendingNames.length ? ", ..." : ""})`
-            : "";
-          const summaryText = summaryParts.join(", ");
-          setLaunchFeedback({
-            status: "warning",
-            message: `Launch in progress: ${summaryText}. Waiting for ${unresolvedPendingCount} confirmation${unresolvedPendingCount === 1 ? "" : "s"}${pendingList}.`,
-          });
-          if (canPollLaunchStatus) {
-            const pollToken = launchStatusPollTokenRef.current;
-            launchStatusPollRef.current = window.setInterval(async () => {
-              if (!isCurrentLaunchContext(launchProfileId, launchRunId, pollToken)) return;
-              try {
-                const statusResult = await window.electron.getLaunchProfileStatus(launchProfileId);
-                if (!isCurrentLaunchContext(launchProfileId, launchRunId, pollToken)) return;
-                const status = statusResult?.status;
-                if (!statusResult?.ok || !status) return;
-                if (String(status.runId || "").trim() !== launchRunId) return;
-                const unresolvedCount = Number(status.unresolvedPendingConfirmationCount || 0);
-                const pollingPendingNames = Array.isArray(status.pendingConfirmations)
-                  ? status.pendingConfirmations
-                    .filter(
-                      (item) => String(item?.status || "waiting").toLowerCase() !== "resolved",
-                    )
-                    .map((item) => String(item?.name || "").trim())
-                    .filter(Boolean)
-                    .slice(0, 3)
-                  : [];
-                if (unresolvedCount <= 0) {
-                  stopLaunchStatusPolling();
-                  activeLaunchRunRef.current = null;
-                  setLaunchFeedback({
-                    status: "success",
-                    message: `Launch complete: ${summaryText}.`,
-                  });
-                  if (launchFeedbackTimeoutRef.current) {
-                    window.clearTimeout(launchFeedbackTimeoutRef.current);
-                  }
-                  launchFeedbackTimeoutRef.current = window.setTimeout(() => {
-                    setLaunchFeedback({
-                      status: "idle",
-                      message: "",
-                    });
-                    launchFeedbackTimeoutRef.current = null;
-                  }, 7000);
-                  return;
-                }
-                const namesList = pollingPendingNames.length > 0
-                  ? ` (${pollingPendingNames.join(", ")}${unresolvedCount > pollingPendingNames.length ? ", ..." : ""})`
-                  : "";
-                setLaunchFeedback({
-                  status: "warning",
-                  message: `Launch in progress: ${summaryText}. Waiting for ${unresolvedCount} confirmation${unresolvedCount === 1 ? "" : "s"}${namesList}.`,
-                });
-              } catch {
-                // Keep UI resilient; next poll tick can recover.
-              }
-            }, 1300);
-          }
-        } else {
-          setLaunchFeedback({
-            status: "success",
-            message: `Launch complete: ${summaryParts.join(", ")}.`,
-          });
-        }
-      } else {
-        activeLaunchRunRef.current = null;
-        const errorMessage = launchResult?.error
-          || "Could not launch this profile. Check app executable paths in app details.";
-        setLaunchFeedback({
-          status: "error",
-          message: errorMessage,
-        });
-        console.error(
-          "Profile launch completed with errors:",
-          launchResult?.error || launchResult?.failedApps || [],
-        );
-      }
-    } catch (error) {
+    void (async () => {
+      stopLaunchStatusPolling();
       activeLaunchRunRef.current = null;
-      console.error("Failed to launch profile:", error);
-      const errorMessage =
-        error instanceof Error && error.message
-          ? error.message
-          : "Launch failed unexpectedly. Please try again.";
+
+      if (launchFeedbackTimeoutRef.current) {
+        window.clearTimeout(launchFeedbackTimeoutRef.current);
+        launchFeedbackTimeoutRef.current = null;
+      }
+
+      setIsLaunching(true);
       setLaunchFeedback({
-        status: "error",
-        message: errorMessage,
+        status: "in-progress",
+        message: "Launching profile...",
       });
-    } finally {
-      setIsLaunching(false);
-      if (!persistLaunchFeedback) {
-        activeLaunchRunRef.current = null;
+
+      let persistLaunchFeedback = false;
+      let deferSpinnerOff = false;
+      const scheduleIdleReset = () => {
+        if (launchFeedbackTimeoutRef.current) {
+          window.clearTimeout(launchFeedbackTimeoutRef.current);
+        }
         launchFeedbackTimeoutRef.current = window.setTimeout(() => {
           setLaunchFeedback({
             status: "idle",
@@ -1509,8 +1360,197 @@ export default function App() {
           });
           launchFeedbackTimeoutRef.current = null;
         }, 7000);
+      };
+      try {
+        if (window.electron?.saveProfiles) {
+          const serializableProfiles = toSerializableProfiles(
+            profiles,
+          );
+          const saveResult = await window.electron.saveProfiles(
+            serializableProfiles,
+          );
+          if (!saveResult?.ok) {
+            setLaunchFeedback({
+              status: "error",
+              message: "Could not save profile changes before launch.",
+            });
+            setIsLaunching(false);
+            launchFeedbackTimeoutRef.current = window.setTimeout(() => {
+              setLaunchFeedback({
+                status: "idle",
+                message: "",
+              });
+              launchFeedbackTimeoutRef.current = null;
+            }, 7000);
+            return;
+          }
+        }
+
+        const launchResult = await window.electron.launchProfile(
+          launchProfileId,
+          { fireAndForget: true },
+        );
+
+        if (!launchResult?.ok) {
+          activeLaunchRunRef.current = null;
+          const errorMessage = launchResult?.error
+            || "Could not launch this profile. Check app executable paths in app details.";
+          setLaunchFeedback({
+            status: "error",
+            message: errorMessage,
+          });
+          console.error(
+            "Profile launch completed with errors:",
+            launchResult?.error || launchResult?.failedApps || [],
+          );
+          return;
+        }
+
+        const launchRunId = String(launchResult?.runId || "").trim();
+        if (!launchResult?.started || !launchRunId) {
+          activeLaunchRunRef.current = null;
+          setLaunchFeedback({
+            status: "error",
+            message: "Launch did not start. Try again.",
+          });
+          return;
+        }
+
+        activeLaunchRunRef.current = {
+          profileId: launchProfileId,
+          runId: launchRunId,
+        };
+        persistLaunchFeedback = true;
+        deferSpinnerOff = true;
+
+        const pollToken = launchStatusPollTokenRef.current;
+        launchStatusPollRef.current = window.setInterval(async () => {
+          if (!isCurrentLaunchContext(launchProfileId, launchRunId, pollToken)) return;
+          try {
+            const statusResult = await window.electron.getLaunchProfileStatus(launchProfileId);
+            if (!isCurrentLaunchContext(launchProfileId, launchRunId, pollToken)) return;
+            const status = statusResult?.status;
+            if (!statusResult?.ok || !status) return;
+            if (String(status.runId || "").trim() !== launchRunId) return;
+
+            const st = String(status.state || "").toLowerCase();
+            const launchedApps = Number(status.launchedAppCount || 0);
+            const launchedTabs = Number(status.launchedTabCount || 0);
+            const failedCount = Number(status.failedAppCount || 0);
+            const skippedCount = Number(status.skippedAppCount || 0);
+            const summaryParts = [
+              `${launchedApps} app${launchedApps === 1 ? "" : "s"}`,
+              `${launchedTabs} tab${launchedTabs === 1 ? "" : "s"}`,
+            ];
+            if (failedCount > 0) summaryParts.push(`${failedCount} failed`);
+            if (skippedCount > 0) summaryParts.push(`${skippedCount} skipped`);
+            const summaryText = summaryParts.join(", ");
+
+            if (st === "cancelled") {
+              stopLaunchStatusPolling();
+              activeLaunchRunRef.current = null;
+              setIsLaunching(false);
+              setLaunchFeedback({
+                status: "warning",
+                message: "Launch cancelled.",
+              });
+              scheduleIdleReset();
+              return;
+            }
+
+            if (st === "failed") {
+              stopLaunchStatusPolling();
+              activeLaunchRunRef.current = null;
+              setIsLaunching(false);
+              setLaunchFeedback({
+                status: "error",
+                message: `Launch finished with failures (${summaryText}).`,
+              });
+              scheduleIdleReset();
+              return;
+            }
+
+            if (st === "complete") {
+              stopLaunchStatusPolling();
+              activeLaunchRunRef.current = null;
+              setIsLaunching(false);
+              setLaunchFeedback({
+                status: "success",
+                message: `Launch complete: ${summaryText}.`,
+              });
+              scheduleIdleReset();
+              return;
+            }
+
+            if (st === "awaiting-confirmations") {
+              const unresolvedCount = Number(status.unresolvedPendingConfirmationCount || 0);
+              const pollingPendingNames = Array.isArray(status.pendingConfirmations)
+                ? status.pendingConfirmations
+                  .filter(
+                    (item) => String(item?.status || "waiting").toLowerCase() !== "resolved",
+                  )
+                  .map((item) => String(item?.name || "").trim())
+                  .filter(Boolean)
+                  .slice(0, 3)
+                : [];
+              if (unresolvedCount > 0) {
+                const namesList = pollingPendingNames.length > 0
+                  ? ` (${pollingPendingNames.join(", ")}${unresolvedCount > pollingPendingNames.length ? ", ..." : ""})`
+                  : "";
+                setLaunchFeedback({
+                  status: "warning",
+                  message: `Launch in progress: ${summaryText}. Waiting for ${unresolvedCount} confirmation${unresolvedCount === 1 ? "" : "s"}${namesList}.`,
+                });
+                return;
+              }
+              stopLaunchStatusPolling();
+              activeLaunchRunRef.current = null;
+              setIsLaunching(false);
+              setLaunchFeedback({
+                status: "success",
+                message: `Launch complete: ${summaryText}.`,
+              });
+              scheduleIdleReset();
+              return;
+            }
+
+            setLaunchFeedback({
+              status: "in-progress",
+              message: "Launching profile...",
+            });
+          } catch {
+            // Keep UI resilient; next poll tick can recover.
+          }
+        }, 1300);
+      } catch (error) {
+        activeLaunchRunRef.current = null;
+        console.error("Failed to launch profile:", error);
+        const errorMessage =
+          error instanceof Error && error.message
+            ? error.message
+            : "Launch failed unexpectedly. Please try again.";
+        setLaunchFeedback({
+          status: "error",
+          message: errorMessage,
+        });
+      } finally {
+        if (!deferSpinnerOff) {
+          setIsLaunching(false);
+        }
+        if (!persistLaunchFeedback) {
+          activeLaunchRunRef.current = null;
+          if (!deferSpinnerOff) {
+            launchFeedbackTimeoutRef.current = window.setTimeout(() => {
+              setLaunchFeedback({
+                status: "idle",
+                message: "",
+              });
+              launchFeedbackTimeoutRef.current = null;
+            }, 7000);
+          }
+        }
       }
-    }
+    })();
   };
 
   const handleDragStart = () => {
@@ -3027,6 +3067,20 @@ export default function App() {
                         </>
                       )}
                     </button>
+                    {(isLaunching
+                      || launchFeedback.status === "in-progress"
+                      || launchFeedback.status === "warning")
+                      && window.electron?.cancelProfileLaunch ? (
+                        <button
+                          type="button"
+                          onClick={() => void handleCancelLaunch()}
+                          disabled={isEditMode}
+                          className="inline-flex items-center justify-center gap-2 rounded-lg px-3 md:px-4 py-2 md:py-2.5 text-xs md:text-sm font-medium transition-all duration-150 ease-out bg-flow-surface/90 border border-flow-border/60 text-flow-text-secondary hover:bg-flow-surface-elevated hover:text-flow-text-primary hover:border-flow-border-accent/50 disabled:opacity-50"
+                        >
+                          <X className="w-4 h-4" />
+                          Cancel launch
+                        </button>
+                      ) : null}
                   </div>
                   {launchFeedback.status !== "idle" && (
                     <div
